update doc and fix bug in unhandled  tx set
