+++ src/main/java/com/walmartlabs/x12/standard/StandardX12Parser.java
@@ -383,12 +383,13 @@ public final class StandardX12Parser implements X12Parser<StandardX12Document> {
             X12TransactionSet txSet = transactionParser.parseTransactionSet(transactionSegments, x12Group);
             if (txSet != null) {
                 x12Group.addTransactionSet(txSet);
+            } else {
+                if (unhandledTransactionSet != null) {
+                    unhandledTransactionSet.unhandledTransactionSet(transactionSegments, x12Group);
+                }
             }
         } else {
             LOGGER.warn("No TransactionSetParser has been registered!");
-            if (unhandledTransactionSet != null) {
-                unhandledTransactionSet.unhandledTransactionSet(transactionSegments, x12Group);
-            }
         }
     }
